Fixed print_tree to print the node and its descendents.
Previously, print_tree only printed the descendents, which was less
useful and not intuitive. Also, TaskPaper objects are now counted as a
node and may be passed to print_tree as an argument. This is a first
step towards enabling parsing of directories containing multiple
taskpaper documents in future releases.